reply end Club start

+++ board/src/main/resources/static/js/read.js
@@ -74,36 +74,64 @@ replyForm.addEventListener('submit', (e) => {
   // 댓글 작성자와 내용 입력값 가져오기
   const replyer = replyForm.querySelector('#replyer');
   const text = replyForm.querySelector('#text');
+  // 수정인 경우에 값이 들어옴
+  const rno = replyForm.querySelector('#rno');
 
   // 댓글 객체 생성
   const reply = {
     bno: bno, // 게시글 번호
     replyer: replyer.value, // 작성자
     text: text.value, // 내용
+    rno: rno.value,
   };
-  // 서버로 댓글 등록 요청 보내기
-  fetch(`/replies/new`, {
-    method: 'post', // POST 방식으로 요청
-    headers: {
-      'content-type': 'application/json', // JSON 형식으로 데이터 전송
-    },
-    body: JSON.stringify(reply), // 댓글 데이터를 JSON 문자열로 변환하여 전송
-  })
-    .then((response) => response.text()) // 응답 텍스트를 가져옴
-    .then((data) => {
-      // 응답 데이터가 있을 경우
-      if (data) {
-        alert(data + ' 번 댓글 등록');
 
-        // replyForm 내용 제거
-        // 댓글 작성자와 내용 입력칸 비우기
-        replyer.value = '';
-        text.value = '';
+  if (!rno.value) {
+    // 새 댓글 등록
+    // 서버로 댓글 등록 요청 보내기
+    fetch(`/replies/new`, {
+      method: 'post', // POST 방식으로 요청
+      headers: {
+        'content-type': 'application/json', // JSON 형식으로 데이터 전송
+      },
+      body: JSON.stringify(reply), // 댓글 데이터를 JSON 문자열로 변환하여 전송
+    })
+      .then((response) => response.text()) // 응답 텍스트를 가져옴
+      .then((data) => {
+        // 응답 데이터가 있을 경우
+        if (data) {
+          alert(data + ' 번 댓글 등록');
 
-        // 댓글 목록 다시 불러오기
-        replyLoaded();
-      }
-    });
+          // replyForm 내용 제거
+          // 댓글 작성자와 내용 입력칸 비우기
+          replyer.value = '';
+          text.value = '';
+
+          // 댓글 목록 다시 불러오기
+          replyLoaded();
+        }
+      });
+  } else {
+    // 댓글 수정
+    fetch(`/replies/${rno.value}`, {
+      method: 'put',
+      headers: {
+        'content-type': 'application/json',
+      },
+      body: JSON.stringify(reply),
+    })
+      .then((response) => response.text())
+      .then((data) => {
+        if (data) {
+          alert('수정성공');
+
+          replyer.value = '';
+          text.value = '';
+          rno.value = '';
+
+          replyLoaded();
+        }
+      });
+  }
 });
 
 // 이벤트 전파 : 자식요소에 일어난 이벤트는 상위 요소로 전달 됨
@@ -115,21 +143,42 @@ replyList.addEventListener('click', (e) => {
   // 실제 이벤트가 일어난 대상은 누구인가? => e.target
   const btn = e.target;
 
+  console.log(btn);
+
   // closest("요소") : 제일 가까운 상위요소 찾기
   // data- 의 값을 가져오는 건 dataset 이용
   const rno = btn.closest('.reply-row').dataset.rno;
   console.log('rno', rno);
 
-  // 해당 댓글 삭제를 서버에 요청
-  fetch(`/replies/${rno}`, {
-    method: 'delete', // DELETE 메서드 사용
-  })
-    .then((response) => response.text()) // 응답 데이터를 텍스트로 변환
-    .then((data) => {
-      if (data == 'success') {
-        // 성공적으로 삭제되면 알림 표시하고 댓글 목록 새로고침
-        alert('댓글 삭제 성공');
-        replyLoaded();
-      }
-    });
+  // 삭제 or 수정 버튼이 눌러졌을 때만 동작
+  // 삭제 or 수정 버튼이 클릭이 되었는지 구분하기
+
+  if (btn.classList.contains('btn-outline-danger')) {
+    // 해당 댓글 삭제를 서버에 요청
+    fetch(`/replies/${rno}`, {
+      method: 'delete', // DELETE 메서드 사용
+    })
+      .then((response) => response.text()) // 응답 데이터를 텍스트로 변환
+      .then((data) => {
+        if (data == 'success') {
+          // 성공적으로 삭제되면 알림 표시하고 댓글 목록 새로고침
+          alert('댓글 삭제 성공');
+          replyLoaded();
+        }
+      });
+  } else if (btn.classList.contains('btn-outline-success')) {
+    // rno 에 해당하는 댓글 가져온 후 댓글 등록 폼에 가져온 내용 보여주기
+
+    fetch(`/replies/${rno}`, {
+      method: 'get',
+    })
+      .then((response) => response.json())
+      .then((data) => {
+        console.log(data);
+
+        replyForm.querySelector('#rno').value = data.rno;
+        replyForm.querySelector('#replyer').value = data.replyer;
+        replyForm.querySelector('#text').value = data.text;
+      });
+  }
 });
